Incorporate review comments from Simon
- use 3 spaces for indent
- use assert.throws for the tests
- move trans and blocktrans tag implementations to their own i18n.js
file
