+++ lib/parser.js
@@ -18,6 +18,7 @@ var Parser = exports.Parser = function(tokens, environment, origin) {
    this.addTags(require('./tag/loop'));
    this.addTags(require('./tag/loader'));
    this.addTags(require('./tag/misc'));
+   this.addTags(require('./tag/i18n'));
    this.addFilters(require('./filter/arrays'));
    this.addFilters(require('./filter/dates'));
    this.addFilters(require('./filter/logic'));
