+++ test/templatedir/trans/templates/page21.html
@@ -1,0 +1,7 @@
+{% blocktrans %}
+This blocktrans has a subtag in it that should cause a syntax error.
+{% block name %}
+This is default text of the illegal subtag
+{% endblock %}
+
+{% endblocktrans %}
